requirements.txt
streamlit
python-dotenv
openai==0.28